Change average_rating to Float

+++ app/src/main/java/models/Show.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.Serializable
 @Parcelize
 data class Show(
     val id: String,
-    val average_rating: Int,
+    val average_rating: Float,
     val description: String,
     val image_url: String,
     val no_of_reviews: Int = 0,
